% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_permutation_importance.R
\name{compute_roc_auc}
\alias{compute_roc_auc}
\title{Compute AUC}
\usage{
compute_roc_auc(y_true, y_pred)
}
\arguments{
\item{y_true}{Observation}

\item{y_pred}{Prediction}
}
\description{
Compute AUC
}
