% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileInput3.R
\name{fileInput3}
\alias{fileInput3}
\title{Alternative version of shiny::fileInput}
\usage{
fileInput3(
  inputId,
  label,
  multiple = FALSE,
  accept = NULL,
  width = NULL,
  buttonLabel = "Browse...",
  placeholder = "No file selected",
  capture = NULL,
  style
)
}
\description{
Alternative version of shiny::fileInput
}
