% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_permutation_importance.R
\name{geneseng_feature_importance}
\alias{geneseng_feature_importance}
\alias{err_loss}
\alias{ratio_loss}
\title{Perform Permutation Feature Importance}
\usage{
geneseng_feature_importance(n, model, X_train, y_train, metric)

err_loss(model, X_train, y_train, metric = c("Accuracy", "AUC", "RMSE", "MAE"))

ratio_loss(model, X_train, y_train, original_loss, metric)
}
\arguments{
\item{n}{Number of replicates}

\item{model}{ML model (from Python libraries)}

\item{X_train}{Features}

\item{y_train}{Target}

\item{metric}{scoring}
}
\description{
Perform Permutation Feature Importance
}
