% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_permutation_importance_h2o.R
\name{geneseng_feature_importance_h2o}
\alias{geneseng_feature_importance_h2o}
\alias{err_loss_h2o}
\alias{ratio_loss_h2o}
\title{Perform Permutation Feature Importance for H2o}
\usage{
geneseng_feature_importance_h2o(n, model, X_train, y_train, metric)

err_loss_h2o(
  model,
  X_train,
  y_train,
  metric = c("Accuracy", "AUC", "RMSE", "MAE")
)

ratio_loss_h2o(model, X_train, y_train, original_loss, metric)
}
\arguments{
\item{n}{Number of replicates}

\item{model}{ML model (from Python libraries)}

\item{X_train}{Features}

\item{y_train}{Target}

\item{metric}{scoring}
}
\description{
Perform Permutation Feature Importance for H2o
}
