% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneseng_run_App.R
\name{geneseng_run_App}
\alias{geneseng_run_App}
\title{Run Geneseng App}
\usage{
geneseng_run_App(port = 3838, host = "0.0.0.0")
}
\arguments{
\item{port}{port}

\item{host}{host}
}
\description{
Run Geneseng App
}
\details{
Solution found at https://stackoverflow.com/questions/37830819/developing-shiny-app-as-a-package-and-deploying-it-to-shiny-server
}
\examples{
library(genesengApp)
geneseng_run_App()


}
