% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_biomarker_file.R
\name{upload_biomarker_file}
\alias{upload_biomarker_file}
\title{Upload file with special JS}
\usage{
upload_biomarker_file(
  inputId = "upload_file",
  label = "Upload File",
  directory = NULL
)
}
\arguments{
\item{inputId}{Id of the input}

\item{label}{label of Btn}

\item{directory}{import folder}

\item{accept}{format of the input file}
}
\description{
Upload file with special JS
}
